web/{{cookiecutter.project_slug}}/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = {extras = ["email"], version = "^2.2.0"}
pydantic-settings = "^2.0.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
alembic = "^1.11.3"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
jinja2 = "^3.1.2"
celery = {extras = ["redis"], version = "^5.3.1"}
emails = "^0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.286"
mypy = "^1.5.1"
types-passlib = "^1.7.7.13"
pytest = "^7.4.2"
httpx = "^0.24.1"

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "celery",
    "emails",
]
ignore_missing_imports = true

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
